SWL-41 progress in MainContainer/types.ts on replacing $TSFixMes with actual types from their reducer files. moved some types declared in file to the types.ts, specifically in newRequestOpenApiSlice.ts
