+++ src/client/toolkit-refactor/slices/newRequestOpenApiSlice.ts
@@ -10,48 +10,7 @@
 import { assertTypeExhaustion } from '../../../helpers';
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { compose } from 'redux';
-import { $NotUsed, $TSFixMe } from '../../../types';
-
-/**
- * Describes a single element within the NewRequestOpenApi type's array of
- * requests.
- */
-export type OpenApiRequest = {
-  request: {
-    id: number;
-  };
-
-  // No idea why headers is an array of single-property objects, instead of just
-  // strings;
-  headers: { name: string }[];
-
-  urls: string[];
-  endpoint: string;
-  reqServers: string[];
-  serverIds: number[];
-  cookies: string;
-  method: string;
-
-  // Below types should probably be strings, but who knows, really? Types united
-  // with $TSFixMe to "temporarily" turn off type-checking (with how this
-  // project tends to go, these types might be stuck like this for months/years)
-  body: string | $TSFixMe;
-  mediaType: string | $TSFixMe;
-  rawType: string | $TSFixMe;
-};
-
-/**
- * Describes an object for keeping track of all OpenAPI request information.
- */
-export type NewRequestOpenApi = {
-  openApiMetadata: {
-    info: Record<string, $TSFixMe>;
-    tags: $TSFixMe[];
-    serverUrls: $TSFixMe[];
-  };
-
-  openApiReqArray: OpenApiRequest[];
-};
+import { $NotUsed, $TSFixMe, NewRequestOpenApi, OpenApiRequest } from '../../../types';
 
 /**
  * An object with info about how a request should be updated? Maybe?
