+++ src/types.ts
@@ -110,13 +110,13 @@ export interface MainContainerProps {
   reqResItemAdded: (reqRes: ReqRes) => any;
   setWarningMessage: (message: ValidationMessage) => any;
   newRequestHeadersSet: NewRequestHeadersSet;
-  newRequestStreamsSet: () => any;
-  fieldsReplaced: () => any;
+  newRequestStreamsSet: newRequestStreamsSet;
+  fieldsReplaced: (requestFields: NewRequestFields) => any;
   newRequestBodySet: NewRequestBodySet;
-  newTestContentSet: () => any;
+  newTestContentSet: (testContent: string) => any;
   newRequestCookiesSet: NewRequestCookiesSet;
   newRequestSSESet: (arg: boolean) => any;
-  openApiRequestsReplaced: () => any;
+  openApiRequestsReplaced: (request: NewRequestOpenApi) => any;
   composerFieldsReset: () => any;
   setWorkspaceActiveTab: (arg: string) => any;
 }
@@ -170,7 +170,7 @@ export type NewRequestHeadersSet = (obj: NewRequestHeaders) => void;
 export type NewRequestCookiesSet = (obj: NewRequestCookies) => void;
 export type NewRequestBodySet = (obj: NewRequestBody) => void;
 export type NewRequestWebRTCSet = (obj: RequestWebRTC) => void;
-
+export type newRequestStreamsSet = (obj: NewRequestStreams) => void;
 /**
  * Defines the type constract for the NewRequestFields state object.
  *
@@ -409,3 +409,43 @@ export interface WorkspaceContainerProps {
   setWorkspace: React.Dispatch<React.SetStateAction<string>>;
 }
 
+/**
+ * Describes an object for keeping track of all OpenAPI request information.
+ */
+export type NewRequestOpenApi = {
+  openApiMetadata: {
+    info: Record<string, $TSFixMe>;
+    tags: $TSFixMe[];
+    serverUrls: $TSFixMe[];
+  };
+
+  openApiReqArray: OpenApiRequest[];
+};
+
+/**
+ * Describes a single element within the NewRequestOpenApi type's array of
+ * requests.
+ */
+export type OpenApiRequest = {
+  request: {
+    id: number;
+  };
+
+  // No idea why headers is an array of single-property objects, instead of just
+  // strings;
+  headers: { name: string }[];
+
+  urls: string[];
+  endpoint: string;
+  reqServers: string[];
+  serverIds: number[];
+  cookies: string;
+  method: string;
+
+  // Below types should probably be strings, but who knows, really? Types united
+  // with $TSFixMe to "temporarily" turn off type-checking (with how this
+  // project tends to go, these types might be stuck like this for months/years)
+  body: string | $TSFixMe;
+  mediaType: string | $TSFixMe;
+  rawType: string | $TSFixMe;
+};
